+++ packages/oas/src/index.ts
@@ -1,8 +1,8 @@
-import definition from './definition';
-import documentation from './documentation';
-import response from './response';
-import router from './router';
-import validation from './validation';
+import definition from './definition.ts';
+import documentation from './documentation.ts';
+import response from './response.ts';
+import router from './router.ts';
+import validation from './validation.ts';
 
 export {
   definition,