+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [v0.0.2] - 2026-02-19
+
+commit [#]()
+
+### Minor changes
+
+- Bumbed dependencies.
+- Removed undici in favor of built-in fetch and jest in favor of built-in test runner.
+
+### Bug fixes
+- Fixed definition middleware checking for compatibility with APM tools.
+
 ## [v0.0.1] - 2025-10-24
 
 commit [#]()